+++ src/main/groovy/fr/nereide/project/OfbizPatterns.groovy
@@ -40,34 +40,6 @@ class OfbizPatterns {
     // =============================================================
     class JAVA {
         public static final PsiElementPattern SERVICE_CALL = psiElement().andOr(
-                psiElement().withTreeParent(literalExpression().methodCallParameter(0, psiMethod()
-                        .withName("makeValidContext")
-                        .definedInClass("org.apache.ofbiz.service.DispatchContext"))),
-
-                psiElement().withTreeParent(literalExpression().methodCallParameter(0, psiMethod()
-                        .withName("runSync")
-                        .definedInClass("org.apache.ofbiz.service.LocalDispatcher"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("runAsync")
-                                .definedInClass("org.apache.ofbiz.service.LocalDispatcher"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("runAsyncWait")
-                                .definedInClass("org.apache.ofbiz.service.LocalDispatcher"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("runSyncIgnore")
-                                .definedInClass("org.apache.ofbiz.service.LocalDispatcher"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("runSyncNewTrans")
-                                .definedInClass("org.apache.ofbiz.service.LocalDispatcher"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("runService")
-                                .definedInClass("org.apache.ofbiz.base.util.ScriptHelper"))),
-
                 literalExpression()
                         .methodCallParameter(0, psiMethod().withName("makeValidContext")
                                 .definedInClass("org.apache.ofbiz.service.DispatchContext")),
@@ -94,42 +66,14 @@ class OfbizPatterns {
 
                 literalExpression()
                         .methodCallParameter(0, psiMethod().withName("runService")
-                                .definedInClass("org.apache.ofbiz.base.util.ScriptHelper"))
+                                .definedInClass("org.apache.ofbiz.base.util.ScriptHelper")),
+
+                literalExpression()
+                        .methodCallParameter(0, psiMethod().withName("schedule")
+                                .definedInClass("org.apache.ofbiz.service.LocalDispatcher")),
         )
 
         public static final PsiElementPattern ENTITY_CALL = psiElement().andOr(
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("find")
-                                .definedInClass("org.apache.ofbiz.entity.Delegator"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("findOne")
-                                .definedInClass("org.apache.ofbiz.entity.Delegator"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("findAll")
-                                .definedInClass("org.apache.ofbiz.entity.Delegator"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("findCountByCondition")
-                                .definedInClass("org.apache.ofbiz.entity.Delegator"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("findList")
-                                .definedInClass("org.apache.ofbiz.entity.Delegator"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(1, psiMethod().withName("addMemberEntity")
-                                .definedInClass("org.apache.ofbiz.entity.model.DynamicViewEntity"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(1, psiMethod().withName("makeGenericValue")
-                                .definedInClass("org.apache.ofbiz.entityext.data.EntityDataServices"))),
-
-                psiElement().withTreeParent(literalExpression()
-                        .methodCallParameter(0, psiMethod().withName("from")
-                                .definedInClass("org.apache.ofbiz.entity.util.EntityQuery"))),
-
                 literalExpression().methodCallParameter(0, psiMethod()
                         .withName("find").definedInClass("org.apache.ofbiz.entity.Delegator")),
 
@@ -152,7 +96,12 @@ class OfbizPatterns {
                         withName("makeGenericValue").definedInClass("org.apache.ofbiz.entityext.data.EntityDataServices")),
 
                 literalExpression().methodCallParameter(0, psiMethod()
-                        .withName("from").definedInClass("org.apache.ofbiz.entity.util.EntityQuery"))
+                        .withName("from").definedInClass("org.apache.ofbiz.entity.util.EntityQuery")),
+
+                literalExpression().methodCallParameter(0,
+                        psiMethod()
+                                .withName("makeValue")
+                                .definedInClass("org.apache.ofbiz.entity.Delegator"))
         )
 
         public static final PsiElementPattern LABEL_CALL = psiElement().andOr(
@@ -219,9 +168,7 @@ class OfbizPatterns {
 
                 groovyLiteralExpression().methodCallParameter(0, psiMethod().withName("from")),
 
-                // TODO : Marche pas
-                groovyLiteralExpression().methodCallParameter(0, psiMethod().withName("makeValue")
-                        .definedInClass("org.apache.ofbiz.service.engine.GroovyBaseScript")),
+                groovyLiteralExpression().methodCallParameter(0, psiMethod().withName("makeValue")),
 
                 groovyLiteralExpression().methodCallParameter(0, psiMethod().withName("findOne"))
         )
@@ -312,6 +259,20 @@ class OfbizPatterns {
                                         .withChild(xmlAttribute()
                                                 .withName("type")
                                                 .withValue(string().contains("service")))
+                                )
+                ),
+                xmlAttributeValue().inside(
+                        xmlAttribute()
+                                .withName('name')
+                                .inside(xmlTag()
+                                        .withName('invoke')
+                                        .inside(xmlTag()
+                                                .withName('group')
+                                                .inside(xmlTag()
+                                                        .withName('service')
+                                                        .withAttributeValue('engine', 'group')
+                                                )
+                                        )
                                 )
                 )
         )