+++ src/main/groovy/fr/nereide/reference/common/EntityReference.groovy
@@ -15,30 +15,24 @@
  *  under the License.
  */
 
-package fr.nereide.reference.xml
-
+package fr.nereide.reference.common
 
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReferenceBase
-import com.intellij.psi.xml.XmlAttributeValue
-import com.intellij.psi.xml.XmlElement
 import com.intellij.util.xml.DomElement
 import fr.nereide.project.ProjectServiceInterface
 import org.jetbrains.annotations.Nullable
 
-class EntityReference extends PsiReferenceBase<XmlElement> {
-    EntityReference(XmlAttributeValue entityName, boolean soft) {
-        super(entityName, soft)
+class EntityReference extends PsiReferenceBase<PsiElement> {
+    EntityReference(PsiElement entityName) {
+        super(entityName)
     }
 
     @Nullable
     PsiElement resolve() {
         ProjectServiceInterface structureService = this.getElement().getProject().getService(ProjectServiceInterface.class)
-
         DomElement definition = structureService.getEntity(this.getValue())
-        // Si on ne trouve pas dans les définitions d'entité, on cherche dans les vues
         if (!definition) definition = structureService.getViewEntity(this.getValue())
-
         return definition != null ? definition.getXmlElement() : null
     }
 }