+++ src/main/groovy/fr/nereide/reference/common/provider/ServiceReferenceProvider.groovy
@@ -15,25 +15,22 @@
  *  under the License.
  */
 
-package fr.nereide.reference.xml
+package fr.nereide.reference.common.provider
 
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
 import com.intellij.psi.PsiReferenceProvider
-import com.intellij.psi.xml.XmlAttributeValue
 import com.intellij.util.ProcessingContext
+import fr.nereide.reference.common.ServiceReference
 import org.jetbrains.annotations.NotNull
 
-class FormReferenceProvider extends PsiReferenceProvider {
-    FormReferenceProvider() {}
+class ServiceReferenceProvider extends PsiReferenceProvider {
+    ServiceReferenceProvider() {}
 
     @NotNull
     PsiReference[] getReferencesByElement(@NotNull PsiElement element, @NotNull ProcessingContext context) {
-        if (element instanceof XmlAttributeValue) {
-            FormReference form = new FormReference((XmlAttributeValue) element, true)
-            PsiReference[] reference = (PsiReference) form
-            return reference
-        }
-        return PsiReference.EMPTY_ARRAY
+        ServiceReference service = new ServiceReference(element)
+        PsiReference[] reference = (PsiReference) service
+        return reference ?: PsiReference.EMPTY_ARRAY
     }
 }