+++ src/main/groovy/fr/nereide/reference/contributor/GroovyReferenceContributor.groovy
@@ -18,49 +18,20 @@
 package fr.nereide.reference.contributor
 
 
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiReference
 import com.intellij.psi.PsiReferenceContributor
-import com.intellij.psi.PsiReferenceProvider
 import com.intellij.psi.PsiReferenceRegistrar
-import com.intellij.util.ProcessingContext
 import fr.nereide.project.OfbizPatterns
-import fr.nereide.reference.groovy.EntityGroovyReference
-import fr.nereide.reference.groovy.ServiceGroovyReference
-import fr.nereide.reference.groovy.UiLabelGroovyReference
+import fr.nereide.reference.common.provider.EntityReferenceProvider
+import fr.nereide.reference.common.provider.ServiceReferenceProvider
+import fr.nereide.reference.common.provider.UiLabelReferenceProvider
 import org.jetbrains.annotations.NotNull
-import org.jetbrains.plugins.groovy.lang.psi.api.statements.expressions.literals.GrLiteral
 
 class GroovyReferenceContributor extends PsiReferenceContributor {
     GroovyReferenceContributor() {}
 
     void registerReferenceProviders(@NotNull PsiReferenceRegistrar registrar) {
-        registrar.registerReferenceProvider(OfbizPatterns.GROOVY.SERVICE_CALL, new PsiReferenceProvider() {
-            @NotNull
-            PsiReference[] getReferencesByElement(@NotNull PsiElement element, @NotNull ProcessingContext context) {
-                GrLiteral el = (GrLiteral) element
-                ServiceGroovyReference service = new ServiceGroovyReference(el, true)
-                PsiReference[] reference = (PsiReference) service
-                return reference
-            }
-        })
-        registrar.registerReferenceProvider(OfbizPatterns.GROOVY.ENTITY_CALL, new PsiReferenceProvider() {
-            @NotNull
-            PsiReference[] getReferencesByElement(@NotNull PsiElement element, @NotNull ProcessingContext context) {
-                GrLiteral el = (GrLiteral) element
-                EntityGroovyReference entity = new EntityGroovyReference(el, true)
-                PsiReference[] reference = (PsiReference) entity
-                return reference
-            }
-        })
-        registrar.registerReferenceProvider(OfbizPatterns.GROOVY.LABEL_CALL, new PsiReferenceProvider() {
-            @NotNull
-            PsiReference[] getReferencesByElement(@NotNull PsiElement element, @NotNull ProcessingContext context) {
-                GrLiteral el = (GrLiteral) element
-                UiLabelGroovyReference label = new UiLabelGroovyReference(el, true)
-                PsiReference[] reference = (PsiReference) label
-                return reference
-            }
-        })
+        registrar.registerReferenceProvider(OfbizPatterns.GROOVY.SERVICE_CALL, new ServiceReferenceProvider())
+        registrar.registerReferenceProvider(OfbizPatterns.GROOVY.ENTITY_CALL, new EntityReferenceProvider())
+        registrar.registerReferenceProvider(OfbizPatterns.GROOVY.LABEL_CALL, new UiLabelReferenceProvider())
     }
 }