+++ src/main/groovy/fr/nereide/reference/xml/provider/RequestMapReferenceProvider.groovy
@@ -15,7 +15,7 @@
  *  under the License.
  */
 
-package fr.nereide.reference.xml
+package fr.nereide.reference.xml.provider
 
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.psi.PsiElement
@@ -23,6 +23,7 @@ import com.intellij.psi.PsiReference
 import com.intellij.psi.PsiReferenceProvider
 import com.intellij.psi.xml.XmlAttributeValue
 import com.intellij.util.ProcessingContext
+import fr.nereide.reference.xml.RequestMapReference
 import org.jetbrains.annotations.NotNull
 
 class RequestMapReferenceProvider extends PsiReferenceProvider {