+++ src/main/groovy/fr/nereide/reference/xml/provider/ViewMapReferenceProvider.groovy
@@ -15,26 +15,28 @@
  *  under the License.
  */
 
-package fr.nereide.reference.xml
+package fr.nereide.reference.xml.provider
 
+import com.intellij.openapi.diagnostic.Logger
 import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiReferenceBase
+import com.intellij.psi.PsiReference
+import com.intellij.psi.PsiReferenceProvider
 import com.intellij.psi.xml.XmlAttributeValue
-import com.intellij.util.xml.DomElement
-import fr.nereide.project.ProjectServiceInterface
-import fr.nereide.project.utils.MiscUtils
-import org.jetbrains.annotations.Nullable
+import com.intellij.util.ProcessingContext
+import fr.nereide.reference.xml.ViewMapReference
+import org.jetbrains.annotations.NotNull
 
-class UiLabelReference extends PsiReferenceBase<XmlAttributeValue> {
-    UiLabelReference(XmlAttributeValue element, boolean soft) {
-        super(element, soft)
+class ViewMapReferenceProvider extends PsiReferenceProvider {
+    ViewMapReferenceProvider() {}
+
+    private static final Logger LOG = Logger.getInstance(ViewMapReferenceProvider.class)
+
+    PsiReference[] getReferencesByElement(@NotNull PsiElement element, @NotNull ProcessingContext context) {
+        if (element instanceof XmlAttributeValue) {
+            ViewMapReference controller = new ViewMapReference((XmlAttributeValue) element, true)
+            PsiReference[] reference = (PsiReference) controller
+            return reference
+        }
+        return PsiReference.EMPTY_ARRAY
     }
-
-    @Nullable
-    PsiElement resolve() {
-        ProjectServiceInterface structureService = this.getElement().getProject().getService(ProjectServiceInterface.class)
-        DomElement definition = structureService.getProperty(MiscUtils.getUiLabelSafeValue(this.getValue()))
-        return definition != null ? definition.getXmlElement() : null
-    }
-
 }