+++ src/test/groovy/fr/nereide/test/reference/GenericRefTestCase.groovy
@@ -18,10 +18,8 @@
 package fr.nereide.test.reference
 
 import com.intellij.lang.properties.references.PropertyReference
-import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiReference
 import com.intellij.psi.impl.source.resolve.reference.impl.PsiMultiReference
-import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase
 import fr.nereide.test.GenericOfbizPluginTestCase
 
 class GenericRefTestCase extends GenericOfbizPluginTestCase {
@@ -37,29 +35,8 @@ class GenericRefTestCase extends GenericOfbizPluginTestCase {
         return "src/test/resources/testData/reference"
     }
 
-    /**
-     * Workaround for groovy refs, which brings back a multi ref object.
-     * We don't want that
-     * @param testFolder
-     * @param type
-     * @return
-     */
-    PsiReference setupFixtureForTestAndGetRefForGroovy(String file) {
-        myFixture.configureByFile(file)
-        PsiReference ref = myFixture.getReferenceAtCaretPositionWithAssertion(file)
-        if (ref instanceof PsiMultiReference) {
-            def multi = ref as PsiMultiReference
-            for (PsiReference curRef : multi.getReferences()) {
-                if (!(curRef instanceof PropertyReference)) {
-                    return curRef
-                }
-            }
-        }
-        return ref
-    }
-
-    protected void configureByFileAndTestRefTypeAndValueForXml(String file, Class expectedRefType, String expectedRefValueName,
-                                                               boolean strict) {
+    protected void configureByFileAndTestRefTypeAndValue(String file, Class expectedRefType, String expectedRefValueName,
+                                                         boolean strict) {
         myFixture.configureByFile(file)
         PsiReference ref = myFixture.getReferenceAtCaretPositionWithAssertion()
         if (ref instanceof PsiMultiReference) {
@@ -67,15 +44,23 @@ class GenericRefTestCase extends GenericOfbizPluginTestCase {
         }
         assert expectedRefType.isAssignableFrom(ref.getClass())
         if (strict) {
-            assertEquals expectedRefValueName, ref.getElement().getName() as String
+            assertEquals expectedRefValueName, ref.getElement().getName() ?
+                    ref.getElement().getName() : getSafeTextInReference(ref)
+                    as String
         } else {
             assert ref.getElement().getText().contains(expectedRefValueName)
         }
         assertNotNull "Reference for $expectedRefValueName not found", ref.resolve()
     }
 
-    protected void configureByFileAndTestRefTypeAndValueForXml(String file, Class expectedRefType, String expectedRefValueName) {
-        configureByFileAndTestRefTypeAndValueForXml(file, expectedRefType, expectedRefValueName, true)
+    static String getSafeTextInReference(PsiReference ref) {
+        ref.getElement().getText()
+                .replaceAll('"', '')
+                .replaceAll('\'', '')
+    }
+
+    protected void configureByFileAndTestRefTypeAndValue(String file, Class expectedRefType, String expectedRefValueName) {
+        configureByFileAndTestRefTypeAndValue(file, expectedRefType, expectedRefValueName, true)
     }
 
     /**